+++ lib/gamelib/hashtabl.c
@@ -29,7 +29,7 @@
 
 /***************************************************************************/
 
-static UDWORD HashTest(int64_t iKey1, int64_t iKey2)
+static UDWORD HashTest(intptr_t iKey1, intptr_t iKey2)
 {
 	return (UDWORD) iKey1 + iKey2;
 }
@@ -180,7 +180,7 @@ hashTable_GetElement( HASHTABLE *psTable )
 
 /***************************************************************************/
 
-static UDWORD hashTable_GetHashKey(HASHTABLE *psTable, int64_t iKey1, int64_t iKey2)
+static UDWORD hashTable_GetHashKey(HASHTABLE *psTable, intptr_t iKey1, intptr_t iKey2)
 {
 	ASSERT( psTable != NULL,
 		"hashTable_GetFirst: hash table pointer invalid\n" );
@@ -191,7 +191,7 @@ static UDWORD hashTable_GetHashKey(HASHTABLE *psTable, int64_t iKey1, int64_t iK
 
 /***************************************************************************/
 
-void hashTable_InsertElement(HASHTABLE *psTable, void *psElement, int64_t iKey1, int64_t iKey2)
+void hashTable_InsertElement(HASHTABLE *psTable, void *psElement, intptr_t iKey1, intptr_t iKey2)
 {
 	UDWORD		udwHashIndex;
 	HASHNODE	*psNode;
@@ -225,7 +225,7 @@ void hashTable_InsertElement(HASHTABLE *psTable, void *psElement, int64_t iKey1,
  */
 /***************************************************************************/
 
-void *hashTable_FindElement(HASHTABLE *psTable, int64_t iKey1, int64_t iKey2)
+void *hashTable_FindElement(HASHTABLE *psTable, intptr_t iKey1, intptr_t iKey2)
 {
 	UDWORD		udwHashIndex;
 	HASHNODE	*psNode;
@@ -294,7 +294,7 @@ hashTable_SetNextNode( HASHTABLE *psTable, BOOL bMoveToNextNode )
 /***************************************************************************/
 
 BOOL
-hashTable_RemoveElement(HASHTABLE *psTable, void *psElement, int64_t iKey1, int64_t iKey2)
+hashTable_RemoveElement(HASHTABLE *psTable, void *psElement, intptr_t iKey1, intptr_t iKey2)
 {
 	UDWORD		udwHashIndex;
 	HASHNODE	*psNode, *psPrev;
